+++ phaneron-plugin-decklink/src/decklink_consumer_config.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct DecklinkConsumerConfiguration {
+    pub device_index: usize,
+}
